pyproject.toml
[project]
name = "fueler-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django==3.2.23",
    "django-ninja>=1.3.0",
    "python-dotenv>=1.0.1",
    "requests==2.31.0",
]
